Safari-611.1.10/LayoutTests/fast/css/named-colors-expected.txt

PASS aliceblue is 240,248,255
PASS antiquewhite is 250,235,215
PASS aqua is 0,255,255
PASS aquamarine is 127,255,212
PASS azure is 240,255,255
PASS beige is 245,245,220
PASS bisque is 255,228,196
PASS black is 0,0,0
PASS blanchedalmond is 255,235,205
PASS blue is 0,0,255
PASS blueviolet is 138,43,226
PASS brown is 165,42,42
PASS burlywood is 222,184,135
PASS cadetblue is 95,158,160
PASS chartreuse is 127,255,0
PASS chocolate is 210,105,30
PASS coral is 255,127,80
PASS cornflowerblue is 100,149,237
PASS cornsilk is 255,248,220
PASS crimson is 220,20,60
PASS cyan is 0,255,255
PASS darkblue is 0,0,139
PASS darkcyan is 0,139,139
PASS darkgoldenrod is 184,134,11
PASS darkgray is 169,169,169
PASS darkgreen is 0,100,0
PASS darkgrey is 169,169,169
PASS darkkhaki is 189,183,107
PASS darkmagenta is 139,0,139
PASS darkolivegreen is 85,107,47
PASS darkorange is 255,140,0
PASS darkorchid is 153,50,204
PASS darkred is 139,0,0
PASS darksalmon is 233,150,122
PASS darkseagreen is 143,188,143
PASS darkslateblue is 72,61,139
PASS darkslategray is 47,79,79
PASS darkslategrey is 47,79,79
PASS darkturquoise is 0,206,209
PASS darkviolet is 148,0,211
PASS deeppink is 255,20,147
PASS deepskyblue is 0,191,255
PASS dimgray is 105,105,105
PASS dimgrey is 105,105,105
PASS dodgerblue is 30,144,255
PASS firebrick is 178,34,34
PASS floralwhite is 255,250,240
PASS forestgreen is 34,139,34
PASS fuchsia is 255,0,255
PASS gainsboro is 220,220,220
PASS ghostwhite is 248,248,255
PASS gold is 255,215,0
PASS goldenrod is 218,165,32
PASS gray is 128,128,128
PASS green is 0,128,0
PASS greenyellow is 173,255,47
PASS grey is 128,128,128
PASS honeydew is 240,255,240
PASS hotpink is 255,105,180
PASS indianred is 205,92,92
PASS indigo is 75,0,130
PASS ivory is 255,255,240
PASS khaki is 240,230,140
PASS lavender is 230,230,250
PASS lavenderblush is 255,240,245
PASS lawngreen is 124,252,0
PASS lemonchiffon is 255,250,205
PASS lightblue is 173,216,230
PASS lightcoral is 240,128,128
PASS lightcyan is 224,255,255
PASS lightgoldenrodyellow is 250,250,210
PASS lightgray is 211,211,211
PASS lightgreen is 144,238,144
PASS lightgrey is 211,211,211
PASS lightpink is 255,182,193
PASS lightsalmon is 255,160,122
PASS lightseagreen is 32,178,170
PASS lightskyblue is 135,206,250
PASS lightslategray is 119,136,153
PASS lightslategrey is 119,136,153
PASS lightsteelblue is 176,196,222
PASS lightyellow is 255,255,224
PASS lime is 0,255,0
PASS limegreen is 50,205,50
PASS linen is 250,240,230
PASS magenta is 255,0,255
PASS maroon is 128,0,0
PASS mediumaquamarine is 102,205,170
PASS mediumblue is 0,0,205
PASS mediumorchid is 186,85,211
PASS mediumpurple is 147,112,219
PASS mediumseagreen is 60,179,113
PASS mediumslateblue is 123,104,238
PASS mediumspringgreen is 0,250,154
PASS mediumturquoise is 72,209,204
PASS mediumvioletred is 199,21,133
PASS midnightblue is 25,25,112
PASS mintcream is 245,255,250
PASS mistyrose is 255,228,225
PASS moccasin is 255,228,181
PASS navajowhite is 255,222,173
PASS navy is 0,0,128
PASS oldlace is 253,245,230
PASS olive is 128,128,0
PASS olivedrab is 107,142,35
PASS orange is 255,165,0
PASS orangered is 255,69,0
PASS orchid is 218,112,214
PASS palegoldenrod is 238,232,170
PASS palegreen is 152,251,152
PASS paleturquoise is 175,238,238
PASS palevioletred is 219,112,147
PASS papayawhip is 255,239,213
PASS peachpuff is 255,218,185
PASS peru is 205,133,63
PASS pink is 255,192,203
PASS plum is 221,160,221
PASS powderblue is 176,224,230
PASS purple is 128,0,128
PASS rebeccapurple is 102,51,153
PASS red is 255,0,0
PASS rosybrown is 188,143,143
PASS royalblue is 65,105,225
PASS saddlebrown is 139,69,19
PASS salmon is 250,128,114
PASS sandybrown is 244,164,96
PASS seagreen is 46,139,87
PASS seashell is 255,245,238
PASS sienna is 160,82,45
PASS silver is 192,192,192
PASS skyblue is 135,206,235
PASS slateblue is 106,90,205
PASS slategray is 112,128,144
PASS slategrey is 112,128,144
PASS snow is 255,250,250
PASS springgreen is 0,255,127
PASS steelblue is 70,130,180
PASS tan is 210,180,140
PASS teal is 0,128,128
PASS thistle is 216,191,216
PASS tomato is 255,99,71
PASS turquoise is 64,224,208
PASS violet is 238,130,238
PASS wheat is 245,222,179
PASS white is 255,255,255
PASS whitesmoke is 245,245,245
PASS yellow is 255,255,0
PASS yellowgreen is 154,205,50
